fixing error checking for 2.6 or later
if someone using the ansible provisioner was on ansible 2.5 or earlier the playbook would fail on aws_caller_facts.  This was because even though there was an assert statement checking for ansible 2.6 or later, the playbook would be pre-processed and error out immediatley on the non-existant module (aws_caller_facts only exists on 2.6 and later).  To combat this, the aws specific checks are moved to a role, and the dynamic method (include_role) was used rather than static method.  Dynamic includes are processed during runtime at the point in which that task is encountered allowsing this to run as intended. This means the assert statement will actually run and give a more meaningful error for folks and provide guidance on how to fix.

The readme was also updated to reflect 2.6 and later with links to the issue explaining why.