cleaning up documentation
this seems like a large update but this has been a long time coming.  I am moving all the sample variable files that define a workshp to a folder called sample_workshops.  I am removing the quick instructions and making sure the normal instructions are less verbose.  I am making usre anything that is a one-time setup (or even a one time explanation) is moved to the docs/setup.md file.