cleaning up some old files
there is a bunch of files (including a workshop index page?) that I had on the master branch for no reason.  I am cleaning up some logic and some HTML website stuff that is not happening in the aws_workshop_login_page (decided to go in a different direction).

The F5 workshop needs a t2.large instance or it will be too slow and not work with the new AS3 exercises that are coming.