Handle ENI/EIP creation for CHKP, update teardown to be VPC centric
Previously we had an issue that would cause the Checkpoint GW
Firewall instance in AWS to hit a race condition on instantiation
because we were trying to attach an extra interface to it after
initial creation. If we could do it fast enough, it would work but
if we weren't fast enough then it wouldn't. Once we got a high
enough student count, we were no longer fast enough and some student
environments were left in a bad network state. This patch fixes
that situation and updates teardown accordingly.

This patch also updates teardown to operate on vpc-id and not tag
data which resolves the situation where if a provision fails before
tags are applied to instances, the teardown will fail because there
will be dangling references to EC2 resources and the
VPC/Subnet/ENI deletions would fail. The one exception to that is
EIPs because EIPs are not searchable by VPC and are still left to be
cleaned up by tag data because we lack an alternative.

Signed-off-by: Adam Miller <admiller@redhat.com>