Minor editing, and removal of Upgrade language from section 3
As the workshop doesn't currently have playbooks for doing upgrades, or
documentation on the process for upgrading from RHEL 7 > 8 yet, I
removed the language saying that it was done in section 3. This
will enable the workshop to be GA'd (v1.0) as it is. This can be
added back after the functionality has been added to the workshop.